Rewrote class methods as functions
