--- bin/config_converter
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-
-import subprocess
-import sys
-
-subprocess.run(["python3", "-m", "config_converter.config_converter"] + sys.argv[1:])
